testlapack: update and move helper functions
